Keep socket in non-blocking mode after duping.
The TCPListener.File() sets the underlying socket to be blocking
(http://git.io/veIh6).  This alters the behaviour of Accept such that
when the listener fd is closed, Accept doesn't return an error until the
next connection comes in.

Setting this back to non-blocking allows this to continue to use the
epoll mechanism meaning that Accept will return an error immediately
when the listener fd is closed.