Prevent leaking file descriptors when reloading.
Previously this would create an extra file descriptor per listener every
time it reloaded. This happened because we copied a set of file
descriptors to pass to the temporary child. These had the close_on_exec
flag cleared so that they could be passed across. However, once the
child had been created, they continued to exist in the parent, and
therefore persisted when it re-exec'd.

To resolve this issue, we now generate a single copy of the file
descriptors to be used both when starting the temporary child, and when
re-execing the parent.