Handle changes in fd blocking mode with Go 1.11
With Go 1.11 the os package no longer puts the descriptor into blocking
mode when being constructed[1]. However, in order to maintain previous
behaviour, the os.File.Fd() call does put it into blocking mode[2]. This
subtle change meant that the code broke because we're calling Fd() to
get the descriptor to set it to non-blocking, and then caling it a
second time which undoes that change. Only calling Fd() once, and
re-using the value prevents this happening.

[1]https://golang.org/doc/go1.11#os
[2]https://git.io/fhtYn