Fix leaking file descriptor test.
This was erroneously asserting that the file descriptors were exactly
the same, and in the same order after the reload. This isn't necessarily
the case as in some cases (eg with Go 1.11 running in a container on
CircleCI), they can be in a different order after the reload.

What we actually care about in this test is that the number of them
hasn't increased, so we can assert on the length instead.